+++ src/lib.rs
@@ -101,6 +101,7 @@ email me at duncan.eddy (at) gmail.com.
 // Re-export commonly used types
 pub use access::*;
 pub use attitude::*;
+pub use clients::{celestrak, spacetrack};
 pub use constants::*;
 pub use coordinates::*;
 pub use datasets::*;
@@ -121,11 +122,11 @@ pub use trajectories::*;
 
 // Module declarations
 pub mod access;
-// Note: celestrak, spacetrack, and ccsds are not glob-re-exported since they use
+// Note: the client modules and ccsds are not glob-re-exported since they use
 // namespaced patterns that should be accessed via brahe::ccsds::*, brahe::celestrak::*, etc.
 pub mod attitude;
 pub mod ccsds;
-pub mod celestrak;
+pub mod clients;
 pub mod constants;
 pub mod coordinates;
 pub mod datasets;
@@ -141,7 +142,6 @@ pub mod orbits;
 pub mod propagators;
 pub mod relative_motion;
 pub mod space_weather;
-pub mod spacetrack;
 pub mod spice;
 pub mod time;
 pub mod trajectories;
